feat(f5): safe cleanup, CI, and performance baseline
D1 – remove dead `pipeline_` member and its include from AnalysisPresenter.h
  (0 uses confirmed before removal).

B6 – zero warnings on beamlab_core / beamlab_domain / beamlab (headless):
  • Fix implicit uint64→double conversion in ApplicationBootstrap.cpp
  • Remove unused static function in AnalysisPipeline.cpp
  • Fix sign-conversion and unused-variable in Geant4CsvImporter.cpp
  • Suppress nodiscard in SimulationEngine.cpp with (void) cast
  Activate -Werror per-target on beamlab_core, beamlab, beamlab_domain.

D3 – reconnect 8 test files to build (114→207 unit tests, 100% green):
  • domain: test_MaterialRegistry (ICRU-90 I=78 eV), test_ParticleRegistry
  • platform: test_EventBus, test_CommandBus, test_PluginHost
  • services: test_QueryCache (fix LRU move-to-front on get),
              test_JobScheduler (fix ThreadPool::waitAll + peak counter)
  Compile-fixes: CommandBus.h (shared_ptr for copyable lambda),
  IPlugin.h (ApplicationContext namespace), StorageManager.cpp (using decls),
  AnalysisOrchestrator (namespace ambiguity + AnalysisRunConfig),
  JobScheduler / IAnalysisEngine / FrameStatisticsPlugin (IStorageBackend type),
  SqliteBackend (mutable stmtCache_ + const prepare/recycle),
  ExporterRegistry (variable shadowing).

Baseline – connect tests/performance/ to build (headless, excl. scene3d_render
  which needs Qt+google-benchmark); add MemoryArena.cpp to beamlab_core sources;
  capture tests/performance/baseline.json (macOS ARM64 Debug).

CI – add .github/workflows/ci.yml: configure + headless build + ctest -L unit
  on ubuntu-22.04; separate physics-regression job for BetheBloch NIST.

+++ src/app/ApplicationBootstrap.cpp
@@ -377,7 +377,7 @@ int ApplicationBootstrap::run(int argc, char** argv)
                       << storage->trajectoryCount() << " trajectories\n";
             std::cout << "[perf] Import: " << importSec << "s  "
                       << "Indices: " << indexSec << "s  "
-                      << "Rate: " << static_cast<int>(imported/importSec) << " samples/s\n";
+                      << "Rate: " << static_cast<int>(static_cast<double>(imported)/importSec) << " samples/s\n";
 
             if (imported > 0) {
                 beamlab::data::AxisFrame ax = {{0,0,0},{0,0,1},{1,0,0},{0,1,0}};
